Ros time

+++ src/APIQuadcopter.hpp
@@ -75,10 +75,10 @@ class APIQuadcopter {
 		int channel;
 		uint32_t colorRange[2];
 		Vector currentSpeedValues[2];
-		uint32_t currentSpeedTimestamps[2];
+		unsigned double currentSpeedTimestamps[2];
 		Vector currentAcceleration;
 		Vector currentPositionValues[2];
-		uint32_t currentPositionTimestamps[2];
+		unsigned double currentPositionTimestamps[2];
 		Vector currentOrientation;
 		
 		float targetSpeed;
